requirements.txt
numpy
Pillow
PyQt6